Refactored ChatDomainImpl instantiation. Move logic to factory
